+++ calicam/gui/3d/capture_volume_visualizer.py
@@ -187,7 +187,7 @@ if __name__ == "__main__":
 
     from calicam import __root__
     from calicam.cameras.camera_array_builder import CameraArrayBuilder
-    from calicam.calibration.capture_volume.point_history import (
+    from calicam.calibration.capture_volume.point_estimates import (
         get_point_history,
     )
     from calicam.calibration.capture_volume.capture_volume import CaptureVolume
