+++ tests/test_camera_array.py
@@ -9,7 +9,7 @@ import pickle
 
 from calicam.cameras.camera_array import CameraArray
 from calicam.cameras.camera_array_builder import CameraArrayBuilder
-from calicam.calibration.capture_volume.point_history import (
+from calicam.calibration.capture_volume.point_estimates import (
     PointEstimates,
     get_point_history,
 )
